requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
fake-useragent==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
tqdm==4.66.1 